add a path check on git path-names
ensure all path-names in a git repository checkout run through `pathchk(1)`
with a portable check option.

this is for most posix systems including a path-name component length
check of 14 (on my system, posix is not implementation specific here).

The posix character limits are such worst-case-low, that if it passes it
has something. If this is not an option, see references for a review of
sources to yours own benefits, pathchk(1) is from GNU coreutils.

If running with `-P` (default), a check for path-names containing space:

    $ git ls-files -z | grep -z ' ' | xargs -0 printf "pathchk-space: %s\n"


pathchk(1) properties:

- all paths in the git repository are construed from the portable character
  set (no C0/C1 and other control/non-printable characters like space).
- no path in the git repository starts with a dash (`-`)

pathchk(1) benefits:

running lax shell tooling on the repository then is likely less prone to
potentially dangerous arg injections making the tree overall more
portable.

this is especially useful for interactive use as it reduces potential
side-effects and increases usability of the repository, quicker turn-
arounds. when in doubt git reset hard first.

this results in an improvement for at-hand review of repositories. shift
left. a base-line for automated builds as well.

NOTE: this linter will fail on repositories not containing any files and
      will give undefined results on dirty/broken repository checkouts.

for references, an extensive review of `pathchk(1)`[1] and file naming
wisdom[2].

contributed to the mega-linter project.

1: https://www.maizure.org/projects/decoded-gnu-coreutils/pathchk.html
2: https://dwheeler.com/essays/fixing-unix-linux-filenames.html