Default DEFAULT_BRANCH to HEAD instead of master.
When DEFAULT_BRANCH isn't specified, diff against origin/HEAD to
determine which files to lint in incremental mode. As before, fetch the
precise ref we require if it isn't present locally. This is a breaking
change for users who don't specify DEFAULT_BRANCH and have origin/HEAD
misconfigured. Deprecate DEFAULT_BRANCH since it is only kept for
backwards compatibility.