Define pre-commit hooks. (#1438)
* Make minor grammatical corrections to README.

* Define pre-commit hooks.

The pre-commit framework allows users to run MegaLinter locally as a Git
pre-commit hook. While it does have built-in support for Node.js hooks,
it doesn't support monorepos, because the package.json must be in the
root directory. While it does have built-in support for Docker hooks,
the user experience of overriding arguments to mega-linter-runner is
nicer than that for the Docker image. Hence, hand-roll system hooks
instead, and leverage npx to execute (and, if necessary, download)
mega-linter-runner.