Don't add literal " character to filenames (#1942)
The extra quotes caused MegaLinter to fail to find the first and last
files in the list on non-Windows platforms, because the first has a
literal double quote character prepended to it, and the last has one
appended to it. Stop passing Windows arguments verbatim to eliminate
the need for the quotes on Windows.