Don't add literal double quote characters to filenames (#1984)
* Remove unnecessary spawnSync options.

The detached option defaults to false, and the cwd option defaults to
the current working directory, so there is no need to specify them.

* Don't retrieve spawnRes properties individually.

This isn't necessary and gives the misleading impression that stdout
and stderr may contain something. They cannot since the Docker image
is run with stdio: "inherit", meaning the stdio and stderr streams are
passed through to/from the mega-linter-runner process and hence aren't
captured.

* Don't add literal " character to filenames (#1942)

The extra quotes caused MegaLinter to fail to find the first and last
files in the list on non-Windows platforms, because the first has a
literal double quote character prepended to it, and the last has one
appended to it. Stop passing Windows arguments verbatim to eliminate
the need for the quotes on Windows.

Co-authored-by: nvuillam <nicolas.vuillamy@gmail.com>