Remove def.ws prefix from aggregate sarifreport
This will replace the default workspace prefix
from when creating the aggregate sarif report output.
(def. megalinter-reports.sarif).

In other words, the artifact uri will be

/root/path/of/source/code

instead of as of now:

DEFAULT_WORKSPACE/root/path/of/source/code.

As the DEFAULT_WORKSPACE might only exist during
runtime, it needs to be removed.

Signed-off-by: Josef Andersson <josef.andersson@gmail.com>