Change reporter text for better UX
To prevent users seeing 'See errors details in' text even when their run is not in an error state, lets change this to 'See detailed report in'.

Also removing the 'artifact' text in the message as there are use cases where artifacts are not published and the link takes the user to the build (and not artifacts) page.