Support mixing remote and local configs in EXTENDS
EXTENDS accepts both absolute URLs and relative file paths to config
files to inherit from. Stop assuming that all relative file paths are
relative to the current workspace. This assumption does not hold when
parsing a config inherited from a different repository. This situation
arises most simply when A inherits from B via an absolute URL, and B
inherits from C via a relative file path. Both inherited config files,
B and C, are in a different repository than A. Make a best effort to
infer the URL of the repository root, and use that to correctly resolve
relative file paths within that repository recursively.