Fix: Gitpod issues #2792 (#2793)
* 🔧 chore(Dockerfile): update NPM_VERSION from 9.7.1 to 9.8.0
The NPM_VERSION variable in the Dockerfile has been updated from 9.7.1 to 9.8.0. This update ensures that the latest version of NPM is used in the Gitpod environment.

* 📦 feat(requirements.txt): add redis package for Python development environment
The redis package is added to the requirements.txt file to include it as a dependency for the Python development environment. This package is necessary for interacting with Redis, a popular in-memory data store, in the application.

* 🔧 chore(requirements.txt): reorder python dependencies in alphabetical order
The python dependencies in the requirements.txt file have been reordered in alphabetical order. This change improves readability and makes it easier to locate specific dependencies, which is particularly useful as the list grows.

* 🔧 chore(.gitpod.yml): update init task to activate virtual environment after bootstrap
The init task in the .gitpod.yml file has been updated to activate the Python virtual environment after the bootstrap process. This change ensures that the correct Python environment is used for subsequent tasks, improving the development setup process in Gitpod.