Fix: Removed curly-brackets from if (closes #3025) (#3029)
* Fix: Removed curly-brackets from if (closes #3025)

A mix of linters fail here, starting with actionlint throwing an
(incorrect) "... is always evaluated to true because extra characters
are around ${{ }}". I believe this is accidentally assuming the
entire statement is curly-bracket wrapped, which it isn't.

Regardless, curly brackets are not required for `if` statements in
the YAML, and can be removed. A best-practice approach would be
changing this to the newer '!cancelled()' flag, which does the same
thing, but this will throw errors when Prettier ultimately changes
the single-quotes to double-quotes and breaks the statement.

As such, the best course of action is to simply remove the curly-
brackets from the existing statement to satisfy Checkov and prevent
linting errors in the default-generated file.

* Added changelog entry

---------

Co-authored-by: Edouard Choinière <27212526+echoix@users.noreply.github.com>