Fix issue with clang-format's config arg
clang-format has a very special syntax for passing a config file as an arg.
You must use --style=:myconfig.file.

To make this work with megalinter I updated the clang descriptor files
to correctly list the config file arg as "--style=:" and modified
the Linter.py file to not add a space to any config file args ending
in ":" as clang-format cannot have a space for its cli syntax here.