Change `golangci-lint` lint mode to `project`
It seems `list_of_files` is not the right choice as default lint
mode for the `golangci-lint` linter, for several reasons.

In such mode, the changed `.go` files are sent in a single call to the
linter.

The first issue with that is that, as the linter
[Quick Start](https://golangci-lint.run/welcome/quick-start/) guide
explicitly states, files must come from the same package.

The second issue is that the linter tries to compile each file via
`types.Checker` and, unless the source file is self-contained and has no
references to things declared in other source files of the codebase, the
compilation will fail, even if the references are towards the same
package.

Such compilation errors are reported by `golangci-lint` and labeled as
`typecheck` (which is not a linter) (see
[why do you have `typecheck` errors?](https://golangci-lint.run/welcome/faq/#why-do-you-have-typecheck-errors)
section of the documentation).

The same problem has also been reported in
[`golangci-lint #1574`](https://github.com/golangci/golangci-lint/issues/1574)
by an user who configured his editor to run the linter on single files.
As commented by a core maintainer
[here](https://github.com/golangci/golangci-lint/issues/1574#issuecomment-804500358)
`typecheck` is not included in the "fast" linters list and, therefore,
cannot be run on single files.