Reorders navigation items & fix linters issues (#5746)
* Reorders navigation items

Moves the "LLM Advisor" section to be after the "Reporters" section in the navigation.

This reordering aims to improve the overall organization and flow of the documentation.

* Improves LLM API key validation messages

Updates the error messages for missing LLM API keys to explicitly state the required environment variable name.

This makes it easier for users to identify and set the correct environment variable when configuring the LLM providers.
Also, formats the provider comparison table in the documentation to improve readability.

* cspell

* Excludes LLM advisor files from jscpd

Excludes LLM advisor files from jscpd to avoid duplicate code findings.

Disables LLM advisor in Bitbucket pipelines for bot PRs to prevent unnecessary analysis.

* Fix mypy issues

* flake8 fixes

* api key

* Disables LLM advisor if no provider is configured

Ensures that the LLM advisor is disabled if no LLM provider is selected
or if none of the required API keys for the supported providers are defined.
This prevents unnecessary processing and potential errors when the LLM
advisor is not properly configured.

* flake8

* Guards against LLM provider errors

Prevents errors when the LLM provider is not available or fails, returning an empty dict instead of None, preventing potential downstream issues.

Adds type hints for better code clarity.

* flake8

* [MegaLinter] Apply linters fixes

* Updates LLM advisor prompt

Clarifies that linters are run by MegaLinter in the LLM advisor's system prompt.
This provides more context to the LLM and improves its ability to analyze and explain linting errors.

* Doc

* fix

* doc

* [MegaLinter] Apply linters fixes

* Fix tests

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>