Removes deprecated markdown-link-check linter
Removes the deprecated `markdown-link-check` linter from MegaLinter.
It can be easily replaced by lychee which is much more performant

The linter has been removed in v9.
This commit removes the linter from the list of linters to be executed,
removes the configuration and documentation files,
removes the linter from the docker images and flavors,
and removes the tests associated with the linter.