Enhances custom flavor generator with linter links (#5873)
* Enhances custom flavor generator with linter links

Improves the custom flavor generator by adding links to the official MegaLinter documentation for each selected linter in the README.md. This makes it easier for users to understand the configuration of their custom flavor. Also populates custom flavor author in the README.

* Uses custom flavor repository URL

Uses the correct repository URL for the custom flavor in the README and generator.

This allows users to navigate to the correct repository when clicking the link in the README.