Use ghcr.io images by default + podman option on mega-linter-runner (#5874)
* init

* Migrates to ghcr.io container registry

Updates the docker image registry from docker://oxsecurity to ghcr.io/oxsecurity.
This ensures better discoverability and management of the MegaLinter images.
A new documentation file listing all the standalone linters and their docker images is added.

* Updates Docker image badge links

Updates the Docker image badge and pull count links in the generated documentation to remove the redundant host prefix from the image name.

This change ensures the badge links point to the correct Docker Hub repository, improving the accuracy and usability of the documentation.

* Updates linters and documentation

Updates linter versions for groovy-lint, java-pmd, repository-trufflehog, and snakemake.

Corrects broken links in documentation for disabled linters and updates powershell CLI executable.

Adds Docker Hub image links to flavor documentation and provides documentation on custom flavors.

* Updates default Docker image location

Updates the default Docker image location to the GitHub Container Registry (ghcr.io).

This change ensures that the correct image is pulled, especially for the default `oxsecurity/megalinter` image, improving image availability and reliability.

* Adds podman support

Enables the use of podman as a container engine, providing an alternative to docker.

Updates the runner to use ghcr.io docker images by default.

Adds a new parameter `--container-engine` to specify either docker or podman as the container runner.

* Adds option to specify container engine

Adds a new `--container-engine` option to allow users to specify the Docker engine (`docker` or `podman`) to be used.
This provides more flexibility and control over the container runtime environment.

* Adds ML_DOCKER_NAME to constants

Adds the ML_DOCKER_NAME constant to the megalinter constants file.
This constant is likely needed for referencing the Docker image name
within the automation scripts or other parts of the codebase.

* Fix syntax error by adding missing semicolon for containerEngine declaration

* Update options

* Initializes containerEngine in constructor

The `containerEngine` property was previously being initialized outside of the constructor, which is not best practice.

This commit updates the code to initialize `containerEngine` within the constructor for better code organization and clarity.

* [MegaLinter] Apply linters fixes

* Validates container engine selection

Removes the class constructor and validates the container engine value directly within the `run` method.

This ensures that the container engine is validated before proceeding with the MegaLinter execution.

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>