Automates linter matrix generation + improve CI performances of renovate workflows (#5895)
* Automates linter matrix generation + improve CI performances of renovate workflows

Moves linter matrix generation to build script, removing manual updates to workflow files.

This change automates the process of creating the linter matrix used in GitHub Actions workflows for building and deploying linter images.
It introduces a script that dynamically generates the matrix based on available linter descriptors, ensuring that only active linters are included.

The generated matrix is saved to a JSON file, which is then used by the workflow definitions.
This eliminates the need to manually update the workflow files whenever linters are added, removed, or disabled, reducing the risk of errors and simplifying maintenance.

* Filters linters matrix based on commit message

Filters the linters matrix based on the commit title or PR title to only run relevant linters.

This improves efficiency by avoiding unnecessary linter runs for dependency updates. It extracts the package name from the commit message (specifically for 'chore(deps)' commits) and only runs linters associated with that package. If no package is found or the commit is not a dependency update, all linters are run as before.

* Refactors matrix assignment in linters workflow

Moves matrix assignment to improve code readability and maintainability.
This change ensures that the matrix variable is defined before its use in the subsequent logic.

* Improves linter filtering in CI workflow

Adds logging to the linter filtering logic in the CI workflow
to aid in debugging and clarifies the behavior when no specific
package is matched or the commit is not a chore(deps) commit,
ensuring all linters are used in such cases.

* Adds debug output for package extraction

Enhances debugging in the linters workflow by adding echo statements.
These statements display the extracted package title and the package
name after extraction, aiding in troubleshooting package identification
issues within the workflow.

* Improves dependency update linter filtering

Modifies the linter filtering logic to use case-insensitive matching
when determining applicable linters for dependency updates. This
ensures that linters are correctly applied even if the package name
in the commit title has inconsistent casing.

* [MegaLinter] Apply linters fixes

* Normalizes package name matching in linter workflow

Ensures case-insensitive matching of package names
when filtering linters, preventing potential misses
due to case discrepancies.

* Improves dependency filtering in linters workflow

Updates the workflow to use case-insensitive matching when filtering linters based on dependency updates.

This ensures that linters are correctly filtered regardless of case variations in package names.

* Removes README from .dockerignore

Excludes README.md from the Docker image, as it's not needed
within the container and reduces the image size.

---------

Co-authored-by: Edouard Choinière <27212526+echoix@users.noreply.github.com>